requirements.txt
torch>=2.1
torchvision
Pillow
numpy
opencv-python
tqdm
